Fix #34

+++ core/src/test/java/eu/bittrade/libs/steemj/SteemJIT.java
@@ -382,7 +382,7 @@ public class SteemJIT extends BaseIntegrationTest {
         assertNotNull("expect properties", properties);
         assertThat("expect head block number", properties.getHeadBlockNumber(), greaterThan(6000000L));
         assertTrue(properties.getHeadBlockId().toString().matches("[0-9a-f]{40}"));
-        assertThat(properties.getHeadBlockId().getHashValue().intValue(), greaterThan(123));
+        assertThat(properties.getHeadBlockId().getHashValue().longValue(), greaterThan(123L));
         assertThat(properties.getHeadBlockId().getNumberFromHash(), greaterThan(123));
         assertThat(properties.getTotalPow(), greaterThan(new BigInteger("123")));
     }
